remove node, move IE and opera fix to it being a self called function
