+++ Gruntfile.js
@@ -19,22 +19,12 @@ module.exports = function(grunt) {
      
       browser: { 
       	options: {
-          banner: '/*! <%= pkg.name %> <%= grunt.template.today("yyyy-mm-dd") %>*/\n/*!©2013 Calvin Metcalf @license MIT https://github.com/calvinmetcalf/communist */\n(function(){\n"use strict";\n',
-          seperator:";\n"
+          banner: '/*! <%= pkg.name %> <%= grunt.template.today("yyyy-mm-dd") %>*/\n/*!©2013 Calvin Metcalf @license MIT https://github.com/calvinmetcalf/communist */\n',
+          seperator:";\n",
+          footer : '})();}'
         },
-        files: {'dist/<%= pkg.name %>.js':['src/promiscuous.js','src/top.browser.js','src/<%= pkg.name %>.js','src/bottom.browser.js']}
-      },
-      node:{
-        files: {'lib/<%= pkg.name %>.js':['src/top.node.js','src/<%= pkg.name %>.js','src/bottom.node.js']}
-    }
-    },
-    simplemocha:{
-      options: {
-        globals: ['console'],
-        ui: "bdd",
-        timeout: 20000
-      },
-      all: { src: 'test/node-test.js' }
+        files: {'dist/<%= pkg.name %>.js':['src/IE.js','src/promiscuous.js','src/<%= pkg.name %>.js']}
+      }
     }
   });
 
@@ -44,8 +34,7 @@ module.exports = function(grunt) {
  grunt.loadNpmTasks('grunt-simple-mocha');
   // Default task(s).
   grunt.registerTask('browser',['concat:browser','uglify:browser']);
-  grunt.registerTask('node',['concat:node']);
-  grunt.registerTask('test', ['concat:node','simplemocha']);
-  grunt.registerTask('default', ['browser','node']);
+
+  grunt.registerTask('default', ['browser']);
 
 };
