--- IE.js
@@ -1,3 +1,0 @@
-self.onmessage=function(e){
-	eval(e.data);	
-}
