+++ src/IE.js
@@ -1,0 +1,7 @@
+if (typeof document === "undefined") {
+	self.onmessage=function(e){
+		eval(e.data);	
+	}
+} else {
+(function(){
+	"use strict";
