--- src/bottom.node.js
@@ -1,1 +1,0 @@
-module.exports = c;
