+++ src/communist.js
@@ -1,3 +1,95 @@
+
+
+//this is mainly so the name shows up when you look at the object in the console
+var Communist = function(){};
+//regex out the importScript call and move it up to the top out of the function.
+function moveImports(string){
+	var script;
+	var match = string.match(/(importScripts\(.*\);)/);
+	if(match){
+		script = match[0].replace(/importScripts\((.*)\);?/,function(a,b){if(b){return "importScripts("+b.split(",").map(function(cc){return '"'+c.makeUrl(cc.slice(1,-1))+'"'})+");\n";}else{return "";}})+string.replace(/(importScripts\(.*\);?)/,"\n");
+	}else{
+		script = string;
+	}
+	return script;
+};
+
+function getPath(){
+	var scripts = document.getElementsByTagName("script");
+        var len = scripts.length;
+        var i = 0;
+        while(i<len){
+            if(/communist(\.min)?\.js/.test(scripts[i].src)){
+               return scripts[i].src;
+            }
+            i++;
+        }
+}
+//accepts an array of strings, joins them, and turns them into a worker.
+function makeWorker(strings){
+	var worker;
+	var script =moveImports(strings.join(""));
+	c.URL = c.URL||window.URL || window.webkitURL;
+	try{
+		worker= new Worker(c.URL.createObjectURL(new Blob([script],{type: "text/javascript"})));	
+	}catch(e){
+		worker = new Worker(getPath());
+		worker.postMessage(script);
+	}finally{
+		return worker;
+	}
+}
+//special case of worker only being called once, instead of sending the data
+//we can bake the data into the worker when we make it.
+function oneOff(fun,data){
+	var promise = c.deferred();
+	var worker = makeWorker(['var _self={};\n_self.fun = ',fun,';\n\
+	_self.cb=function(data,transfer){\n\
+			self.postMessage(data,transfer);\n\
+			self.close();\n\
+		};\n\
+		_self.result = _self.fun(',JSON.stringify(data),',_self.cb);\n\
+		if(typeof _self.result !== "undefined"){\n\
+			_self.cb(_self.result);\n\
+		}']);
+	worker.onmessage=function(e){
+		promise.resolve(e.data);
+	};
+	worker.onerror=function(e){
+		e.preventDefault();
+		promise.reject(e.message);
+	};
+	return promise.promise;
+};
+function mapWorker(fun,callback,onerr){
+	var w = new Communist();
+	var worker = makeWorker(['var _close=function(){self.close();};var _db={};\nvar _self={};\n_self.fun = ',fun,';\n\
+		_self.cb=function(data,transfer){\n\
+			self.postMessage(data,transfer);\n\
+		};\n\
+		self.onmessage=function(e){\n\
+		_self.result = _self.fun(e.data,_self.cb);\n\
+			if(typeof _self.result !== "undefined"){\n\
+				_self.cb(_self.result);\n\
+			}\n\
+		}']);
+	worker.onmessage = function(e){
+		callback(e.data);	
+	};
+	if(onerr){
+		worker.onerror=onerr;
+	}else{
+		worker.onerror=function(){callback();};
+	}
+	w.data=function(d,t){
+		worker.postMessage(d,t);	
+		return w;
+	};
+	w.close=function(){
+		return worker.terminate();
+	};
+	return w;
+};
 function sticksAround(fun){
 	var w = new Communist();
 	var promises = [];
@@ -34,7 +126,7 @@ function sticksAround(fun){
 	};
 	w.data=function(data, transfer){
 		var i = promises.length;
-		promises[i] = makePromise();
+		promises[i] = c.deferred();
 		worker.data([i,data],transfer);
 		return promises[i].promise;
 	};
@@ -177,7 +269,7 @@ function incrementalMapReduce(threads){
 	}
 	w.fetch=function(now){
 		if(!promise){
-			promise = makePromise();
+			promise = c.deferred();
 		}
 		if(idle<threads && !now){
 			waiting=true;
@@ -188,7 +280,7 @@ function incrementalMapReduce(threads){
 	};
 	w.close=function(){
 		if(!promise){
-			promise = makePromise();
+			promise = c.deferred();
 		}
 		if(idle<threads){
 			closing=true;
@@ -287,3 +379,26 @@ function c(a,b,c){
 };
 c.reducer = rWorker;
 c.worker = makeWorker;
+c.makeUrl = function (fileName) {
+	var link = document.createElement("link");
+	link.href = fileName;
+	return link.href;
+};
+c.ajax = function(url,after,notjson){
+	var txt=!notjson?'JSON.parse(request.responseText)':"request.responseText";
+	var resp = after?"("+after.toString()+")("+txt+",_cb)":txt;
+	var func = 'function (url, _cb) {\n\
+		var request = new XMLHttpRequest();\n\
+		request.open("GET", url);\n\
+			request.onreadystatechange = function() {\n\
+				var _resp;\n\
+				if (request.readyState === 4 && request.status === 200) {\n'+
+					'_resp = '+resp+';\n\
+					if(typeof _resp!=="undefined"){_cb(_resp);}\n\
+					}\n\
+			};\n\
+		request.send();\n\
+	}';
+	return c(func,c.makeUrl(url));
+};
+window["communist"]=c;
